+++ haskell/ngx-export-distribution/NgxExport/Distribution.hs
@@ -303,7 +303,8 @@ import Data.Maybe
 --
 -- > $ make clean
 --
--- or, to additionally delete the GHC environment file and the built artifact,
+-- or, to additionally delete the GHC environment file, the Cabal build
+-- directory, and the built artifact,
 --
 -- > $ make clean-all
 --